Added code to support persisting of LDAP users
Ldap users are persisted on Logon into generis. 

This allows for results etc to render correctly. 

A minor patch is recommended to the user class to support the creation of "invalid" passwords to ensure that regardless of the authentication order the local logon will never be used (as it will always fail to authenticate)